@chart
This is a small and possibly silly question, but I can't seem to figure out how to change the description of a project on Tableau Server once it has been created.  I figured out how to change the name, but not the description.  My IT Admin created a project for our first live extract, and I'd love to edit the default description he left to be more illustrative of the contents.  The only workaround I can think of is to create a brand new project and then move all of the content over, but it seems there ought to be an easier way!